Adds a singleton for the PIT and hides it behind a timer api
