+++ diy-os/src/lib.rs
@@ -9,6 +9,7 @@
 #![feature(ascii_char)]
 #![feature(ascii_char_variants)]
 #![feature(anonymous_lifetime_in_impl_trait)]
+#![feature(const_refs_to_static)]
 #![warn(clippy::pedantic, clippy::nursery, clippy::perf, clippy::style)]
 #![deny(
     clippy::suspicious,
@@ -39,6 +40,7 @@ pub mod memory;
 pub mod pit;
 pub mod serial;
 pub mod spinlock;
+pub mod timer;
 
 pub trait Testable {
     #[allow(clippy::unused_unit)]
@@ -124,6 +126,8 @@ pub fn init(boot_info: &'static mut BootInfo) -> &'static mut BootInfo {
     println!("Interrupts Unmasked");
     x86_64::instructions::interrupts::enable();
     println!("Interrupts Enabled");
+    timer::setup_system_timer();
+    println!("System timer Initialized");
 
     boot_info
 }
