+++ diy-os/src/main.rs
@@ -2,7 +2,6 @@
 #![no_main]
 #![feature(optimize_attribute)]
 #![feature(custom_test_frameworks)]
-#![feature(f16)]
 #![test_runner(diy_os::test_runner)]
 #![reexport_test_harness_main = "test_main"]
 #![warn(clippy::pedantic, clippy::nursery, clippy::perf, clippy::style)]
@@ -24,10 +23,8 @@ use core::panic::PanicInfo;
 use diy_os::{
     allocator, hlt_loop, init,
     memory::{self, BootInfoFrameAllocator},
-    pit::{self, AccessMode, Pit, ReadBackCommand},
-    println,
+    println, timer,
 };
-use x86_64::instructions::hlt;
 
 static BOOTLOADER_CONFIG: BootloaderConfig = {
     let mut config = BootloaderConfig::new_default();
@@ -54,61 +51,13 @@ extern "Rust" fn main(mut boot_info: &'static mut BootInfo) -> ! {
 
     println!("Hello, world!");
 
-    let mut pit = pit::Pit::new();
-    let read_back_command = pit::ReadBackCommandBuilder::new()
-        .set_read_from_channel_0(true)
-        .set_read_status_byte(true)
-        .build();
-
-    let configure_command = pit::ConfigureChannelCommand::new(
-        pit::Channel::Channel0,
-        AccessMode::LowHighbyte,
-        pit::OperatingMode::SquareWaveGenerator,
-        pit::BcdBinaryMode::Binary16Bit,
-    );
-
-    pit.mode_port.write(configure_command);
-
-    // pit.mode_port.write(read_back_command);
-    //
-    let reaload_value = get_reload_value_from_frequency(1000);
-
-    println!("readlaod vlaue {}", reaload_value);
-
-    set_count(&mut pit, reaload_value);
-
     println!("going to sleep");
 
-    sleep(1000);
+    timer::sleep(1000);
 
     println!("wakign up");
 
     hlt_loop();
-}
-
-pub fn get_reload_value_from_frequency(frequency: u32) -> u16 {
-    u16::try_from(1192182 / frequency).unwrap()
-}
-
-pub fn set_count(pit: &mut Pit, count: u16) -> &mut Pit {
-    x86_64::instructions::interrupts::without_interrupts(|| {
-        pit.channel_0_port.write((count & 0xFF).try_into().unwrap()); // low_byte
-        pit.channel_0_port
-            .write(((count & 0xFF00) >> 8).try_into().unwrap()) // high byte
-    });
-
-    pit
-}
-
-// time in ms
-pub fn sleep(count: u64) {
-    *pit::SLEEP_COUNTER.acquire() = count;
-    pit::SLEEP_COUNTER.release();
-    
-    while *pit::SLEEP_COUNTER.acquire() > 0 {
-        pit::SLEEP_COUNTER.release();
-        hlt();
-    }
 }
 
 /// This function is called on panic.
